+++ src/helpers/constants/Role.js
@@ -4,7 +4,7 @@ exports.groupNames = [...this.sideNames, "lovers", "charmed"];
 
 exports.roles = [
     { name: "werewolf", side: "werewolves", maxInGame: 4 },
-    { name: "big-bad-wolf", side: "werewolves", maxInGame: 1, recommendedMinPlayers: 12 },
+    { name: "big-bad-wolf", side: "werewolves", maxInGame: 1, recommendedMinPlayers: 15 },
     { name: "vile-father-of-wolves", side: "werewolves", maxInGame: 1, recommendedMinPlayers: 12 },
     { name: "villager", side: "villagers", maxInGame: 39 },
     { name: "villager-villager", side: "villagers", maxInGame: 1 },